#75 refactor tags API call for repository editing
